[Web] Refactor the profile layout folder
