+++ core/datacap-web/console-fe/src/views/layout/user-profile/components/ProfileLayoutContent.vue
