+++ core/datacap-web/console-fe/src/views/layout/user-profile/components/ProfileLayoutSidebar.vue
